refactor(editor): extract map editor component
